+++ src/apis/auth/discord.py
@@ -140,6 +140,7 @@ async def get_callback(request: Request, response: Response, code: Optional[str]
                 elif status == 0:
                     await app.db.execute(dhrid, f"DELETE FROM pending_user_deletion WHERE uid = {uid}")
                     await app.db.commit(dhrid)
+                    await AuditLog(request, uid, "user", ml.ctr(request, "cancelled_user_deletion", var = {"username": username, "uid": uid}))
 
             stoken = str(uuid.uuid4())
             while stoken[0] == "e":